+++ app/src/main/kotlin/com/letsdoit/app/di/NetworkModule.kt
@@ -3,6 +3,7 @@ package com.letsdoit.app.di
 import com.letsdoit.app.network.ClickUpService
 import com.letsdoit.app.security.SecurePrefs
 import com.squareup.moshi.Moshi
+import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -58,5 +59,7 @@ object NetworkModule {
 
     @Provides
     @Singleton
-    fun provideMoshi(): Moshi = Moshi.Builder().build()
+    fun provideMoshi(): Moshi = Moshi.Builder()
+        .addLast(KotlinJsonAdapterFactory())
+        .build()
 }